common move to type_guard
