+++ src/libs/utils/type_guard/common.ts
@@ -1,0 +1,4 @@
+// Info Murky (20240505): type guards
+export function isStringNumber(value: any): value is string {
+  return typeof value === 'string' && !isNaN(Number(value));
+}
